restructured code for readability
