+++ backend/.env
@@ -1,3 +1,3 @@
-SPOTIFY_CLIENT_ID=REMOVED
-SPOTIFY_CLIENT_SECRET=REMOVED
-SPOTIFY_REDIRECT_URI=https://a554-185-70-53-13.ngrok-free.app/callback
+SPOTIFY_CLIENT_ID=91d28fed2a0a4698847b26ffa20e676b
+SPOTIFY_CLIENT_SECRET=44fc679e2281441b95b6c56b258b01ce
+SPOTIFY_REDIRECT_URI=https://1bb4-109-36-229-20.ngrok-free.app/callback
