+++ backend/redis_client.py
@@ -1,0 +1,5 @@
+import redis
+
+from config import REDIS_HOST, REDIS_PORT
+
+r = redis.Redis(host=REDIS_HOST, port=REDIS_PORT, decode_responses=True)
