+++ backend/requirements.txt
@@ -5,3 +5,4 @@ beautifulsoup4
 redis
 fastapi[standard]
 uvicorn
+dotenv
