+++ backend/routes/auth.py
@@ -1,0 +1,41 @@
+from fastapi import APIRouter, Request
+from fastapi.responses import RedirectResponse, JSONResponse
+import requests
+from config import SPOTIFY_CLIENT_ID, SPOTIFY_CLIENT_SECRET, SPOTIFY_REDIRECT_URI
+
+router = APIRouter()
+
+@router.get("/login")
+def login():
+    scopes = "user-read-private user-read-email user-follow-read"
+    url = (
+        "https://accounts.spotify.com/authorize"
+        f"?response_type=code&client_id={SPOTIFY_CLIENT_ID}"
+        f"&scope={scopes}&redirect_uri={SPOTIFY_REDIRECT_URI}"
+    )
+    return RedirectResponse(url)
+
+@router.get("/callback")
+def callback(request: Request):
+    code = request.query_params.get("code")
+    if not code:
+        return JSONResponse({"error": "No code provided"}, status_code=400)
+
+    token_url = "https://accounts.spotify.com/api/token"
+    data = {
+        "grant_type": "authorization_code",
+        "code": code,
+        "redirect_uri": SPOTIFY_REDIRECT_URI,
+        "client_id": SPOTIFY_CLIENT_ID,
+        "client_secret": SPOTIFY_CLIENT_SECRET,
+    }
+    headers = {"Content-Type": "application/x-www-form-urlencoded"}
+
+    token_res = requests.post(token_url, data=data, headers=headers)
+    token_json = token_res.json()
+    access_token = token_json.get("access_token")
+
+    if not access_token:
+        return JSONResponse({"error": "Failed to get access token"}, status_code=400)
+
+    return RedirectResponse(f"http://localhost:5173/?token={access_token}")
